+++ src/EntityGraphQL/Schema/FieldToResolve.cs
@@ -70,30 +70,35 @@ public class FieldToResolveWithArgs<TContext, TParams> : FieldWithContextAndArgs
         SetUpField(fieldExpression, true, true);
         return this;
     }
+    public FieldWithContextAndArgs<TContext, TParams> Resolve<TService>(Expression<Func<TContext, TParams, TService, object>> fieldExpression) => ResolveWithService(fieldExpression);
     public FieldWithContextAndArgs<TContext, TParams> ResolveWithService<TService>(Expression<Func<TContext, TParams, TService, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, true);
         Services = new List<ParameterExpression> { fieldExpression.Parameters[2] };
         return this;
     }
+    public FieldWithContextAndArgs<TContext, TParams> Resolve<TService1, TService2>(Expression<Func<TContext, TParams, TService1, TService2, object>> fieldExpression) => ResolveWithServices(fieldExpression);
     public FieldWithContextAndArgs<TContext, TParams> ResolveWithServices<TService1, TService2>(Expression<Func<TContext, TParams, TService1, TService2, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, true);
         Services = new List<ParameterExpression> { fieldExpression.Parameters[2], fieldExpression.Parameters[3] };
         return this;
     }
+    public FieldWithContextAndArgs<TContext, TParams> Resolve<TService1, TService2, TService3>(Expression<Func<TContext, TParams, TService1, TService2, TService3, object>> fieldExpression) => ResolveWithServices(fieldExpression);
     public FieldWithContextAndArgs<TContext, TParams> ResolveWithServices<TService1, TService2, TService3>(Expression<Func<TContext, TParams, TService1, TService2, TService3, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, true);
         Services = new List<ParameterExpression> { fieldExpression.Parameters[2], fieldExpression.Parameters[3], fieldExpression.Parameters[4] };
         return this;
     }
+    public FieldWithContextAndArgs<TContext, TParams> Resolve<TService1, TService2, TService3, TService4>(Expression<Func<TContext, TParams, TService1, TService2, TService3, TService4, object>> fieldExpression) => ResolveWithServices(fieldExpression);
     public FieldWithContextAndArgs<TContext, TParams> ResolveWithServices<TService1, TService2, TService3, TService4>(Expression<Func<TContext, TParams, TService1, TService2, TService3, TService4, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, true);
         Services = new List<ParameterExpression> { fieldExpression.Parameters[2], fieldExpression.Parameters[3], fieldExpression.Parameters[4], fieldExpression.Parameters[5] };
         return this;
     }
+    public FieldWithContextAndArgs<TContext, TParams> Resolve<TService1, TService2, TService3, TService4, TService5>(Expression<Func<TContext, TParams, TService1, TService2, TService3, TService4, TService5, object>> fieldExpression) => ResolveWithServices(fieldExpression);
     public FieldWithContextAndArgs<TContext, TParams> ResolveWithServices<TService1, TService2, TService3, TService4, TService5>(Expression<Func<TContext, TParams, TService1, TService2, TService3, TService4, TService5, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, true);
@@ -118,30 +123,35 @@ public class FieldToResolve<TContext> : FieldWithContext<TContext>
         return this;
     }
 
+    public FieldWithContext<TContext> Resolve<TService>(Expression<Func<TContext, TService, object>> fieldExpression) => ResolveWithService(fieldExpression);
     public FieldWithContext<TContext> ResolveWithService<TService>(Expression<Func<TContext, TService, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, false);
         Services = new List<ParameterExpression> { fieldExpression.Parameters[1] };
         return this;
     }
+    public FieldWithContext<TContext> Resolve<TService1, TService2>(Expression<Func<TContext, TService1, TService2, object>> fieldExpression) => ResolveWithServices(fieldExpression);
     public FieldWithContext<TContext> ResolveWithServices<TService1, TService2>(Expression<Func<TContext, TService1, TService2, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, false);
         Services = new List<ParameterExpression> { fieldExpression.Parameters[1], fieldExpression.Parameters[2] };
         return this;
     }
+    public FieldWithContext<TContext> Resolve<TService1, TService2, TService3>(Expression<Func<TContext, TService1, TService2, TService3, object>> fieldExpression) => ResolveWithServices(fieldExpression);
     public FieldWithContext<TContext> ResolveWithServices<TService1, TService2, TService3>(Expression<Func<TContext, TService1, TService2, TService3, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, false);
         Services = new List<ParameterExpression> { fieldExpression.Parameters[1], fieldExpression.Parameters[2], fieldExpression.Parameters[3] };
         return this;
     }
+    public FieldWithContext<TContext> Resolve<TService1, TService2, TService3, TService4>(Expression<Func<TContext, TService1, TService2, TService3, TService4, object>> fieldExpression) => ResolveWithServices(fieldExpression);
     public FieldWithContext<TContext> ResolveWithServices<TService1, TService2, TService3, TService4>(Expression<Func<TContext, TService1, TService2, TService3, TService4, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, false);
         Services = new List<ParameterExpression> { fieldExpression.Parameters[1], fieldExpression.Parameters[2], fieldExpression.Parameters[3], fieldExpression.Parameters[4] };
         return this;
     }
+    public FieldWithContext<TContext> Resolve<TService1, TService2, TService3, TService4, TService5>(Expression<Func<TContext, TService1, TService2, TService3, TService4, TService5, object>> fieldExpression) => ResolveWithServices(fieldExpression);
     public FieldWithContext<TContext> ResolveWithServices<TService1, TService2, TService3, TService4, TService5>(Expression<Func<TContext, TService1, TService2, TService3, TService4, TService5, object>> fieldExpression)
     {
         SetUpField(fieldExpression, true, false);