+++ src/tests/EntityGraphQL.Tests/QueryTests/ServiceFieldBulkTests.cs
@@ -15,7 +15,7 @@ public class ServiceFieldBulkTests
         schema.UpdateType<Project>(type =>
         {
             type.ReplaceField("createdBy", "Get user that created it")
-                .ResolveWithService<UserService>((proj, users) => users.GetUserById(proj.CreatedBy))
+                .Resolve<UserService>((proj, users) => users.GetUserById(proj.CreatedBy))
                 .ResolveBulk<UserService, int, User>(proj => proj.CreatedBy, (ids, srv) => srv.GetAllUsers(ids));
         });
 
@@ -61,7 +61,7 @@ public class ServiceFieldBulkTests
         schema.UpdateType<Project>(type =>
         {
             type.AddField("createdByName", "Get user name that created it")
-                .ResolveWithService<UserService>((proj, users) => users.GetUserById(proj.CreatedBy).Name)
+                .Resolve<UserService>((proj, users) => users.GetUserById(proj.CreatedBy).Name)
                 .ResolveBulk<UserService, int, string>(proj => proj.CreatedBy, (ids, srv) => srv.GetAllUsers(ids).ToDictionary(u => u.Key, u => u.Value.Name));
         });
 
@@ -105,7 +105,7 @@ public class ServiceFieldBulkTests
         schema.UpdateType<Project>(type =>
         {
             type.AddField("assignedUsers", "Get users assigned to project")
-                .ResolveWithService<UserService>((proj, users) => users.GetUsersByProjectId(proj.Id, null))
+                .Resolve<UserService>((proj, users) => users.GetUsersByProjectId(proj.Id, null))
                 .ResolveBulk<UserService, int, List<User>>(proj => proj.Id, (ids, srv) => srv.GetUsersByProjectId(ids, null));
         });
 
@@ -149,7 +149,7 @@ public class ServiceFieldBulkTests
         schema.UpdateType<Project>(type =>
         {
             type.AddField("assignedUser", new { id = ArgumentHelper.Required<int>() }, "Get user assigned to project by ID")
-                .ResolveWithService<UserService>((proj, args, users) => users.GetUserByProjectId(proj.Id, args.id))
+                .Resolve<UserService>((proj, args, users) => users.GetUserByProjectId(proj.Id, args.id))
                 .ResolveBulk<UserService, int, User>(proj => proj.Id, (ids, args, srv) => srv.GetUserByIdForProjectId(ids, args.id));
         });
 
@@ -193,7 +193,7 @@ public class ServiceFieldBulkTests
         schema.UpdateType<Project>(type =>
         {
             type.AddField("assignedUsers", new { name = (string)null }, "Get users assigned to project")
-                .ResolveWithService<UserService>((proj, args, users) => users.GetUsersByProjectId(proj.Id, args.name))
+                .Resolve<UserService>((proj, args, users) => users.GetUsersByProjectId(proj.Id, args.name))
                 .ResolveBulk<UserService, int, List<User>>(proj => proj.Id, (ids, args, srv) => srv.GetUsersByProjectId(ids, args.name));
         });